# task11/export/8uopx2p5ga_oas_v3.json
{
   "openapi":"3.0.1",
   "info":{
      "title":"task11_api",
      "version":"2024-06-19T09:44:38Z"
   },
   "servers":[
      {
         "url":"https://v0rqgq9710.execute-api.eu-central-1.amazonaws.com/{basePath}",
         "variables":{
            "basePath":{
               "default":"api"
            }
         }
      }
   ],
...
   "paths":{
      "/reservations":{
         "get":{
            "tags":[
               "Reservations"
            ],
            "summary":"Operations with reservations",
            "description":"Operations with reservations",
            "responses":{
               "200":{
                  "description":"200 response",
                  "headers":{
                     "Access-Control-Allow-Origin":{
                        "schema":{
                           "type":"string"
                        }
                     }
                  },
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/ReservationsList"
                        }
                     }
                  }
               }
            },
            "security":[
               {
                  "authorizer":[

                  ]
               }
            ],
            "x-amazon-apigateway-integration":{
               "httpMethod":"POST",
               "uri":"arn:aws:apigateway:eu-central-1:lambda:path/2015-03-31/functions/arn:aws:lambda:eu-central-1:196241772369:function:api_handler/invocations",
               "responses":{
                  "default":{
                     "statusCode":"200",
                     "responseParameters":{
                        "method.response.header.Access-Control-Allow-Origin":"'*'"
                     }
                  }
               },
               "passthroughBehavior":"when_no_match",
               "type":"aws_proxy"
            }
         },
         "post":{
            "tags":[
               "Reservations"
            ],
            "summary":"Operations with reservations",
            "description":"Operations with reservations",
            "requestBody":{
               "description":"Create a new pet in the store",
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/Reservation"
                     }
                  }
               }
            },
            "responses":{
               "200":{
                  "description":"200 response",
                  "headers":{
                     "Access-Control-Allow-Origin":{
                        "schema":{
                           "type":"string"
                        }
                     }
                  },
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/ReservationId"
                        }
                     }
                  }
               }
            },
            "security":[
               {
                  "authorizer":[

                  ]
               }
            ],
            "x-amazon-apigateway-integration":{
               "httpMethod":"POST",
               "uri":"arn:aws:apigateway:eu-central-1:lambda:path/2015-03-31/functions/arn:aws:lambda:eu-central-1:196241772369:function:api_handler/invocations",
               "responses":{
                  "default":{
                     "statusCode":"200",
                     "responseParameters":{
                        "method.response.header.Access-Control-Allow-Origin":"'*'"
                     }
                  }
               },
               "passthroughBehavior":"when_no_match",
               "type":"aws_proxy"
            }
         },
         "options":{
            "tags":[
               "Reservations"
            ],
            "responses":{
               "200":{
                  "description":"200 response",
                  "headers":{
                     "Access-Control-Allow-Origin":{
                        "schema":{
                           "type":"string"
                        }
                     },
                     "Access-Control-Allow-Methods":{
                        "schema":{
                           "type":"string"
                        }
                     },
                     "Access-Control-Allow-Headers":{
                        "schema":{
                           "type":"string"
                        }
                     }
                  },
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/Empty"
                        }
                     }
                  }
               }
            },
            "x-amazon-apigateway-integration":{
               "responses":{
                  "default":{
                     "statusCode":"200",
                     "responseParameters":{
                        "method.response.header.Access-Control-Allow-Methods":"'*'",
                        "method.response.header.Access-Control-Allow-Headers":"'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token'",
                        "method.response.header.Access-Control-Allow-Origin":"'*'"
                     }
                  }
               },
               "requestTemplates":{
                  "application/json":"{\"statusCode\": 200}"
               },
               "passthroughBehavior":"when_no_match",
               "type":"mock"
            }
         }
      },
      "/tables":{
         "get":{
            "summary":"Operations with tables",
            "description":"Operations with tables",
            "tags":[
               "Tables"
            ],
            "responses":{
               "200":{
                  "description":"200 response",
                  "headers":{
                     "Access-Control-Allow-Origin":{
                        "schema":{
                           "type":"string"
                        }
                     }
                  },
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/TableList"
                        }
                     }
                  }
               }
            },
            "security":[
               {
                  "authorizer":[

                  ]
               }
            ],
            "x-amazon-apigateway-integration":{
               "httpMethod":"POST",
               "uri":"arn:aws:apigateway:eu-central-1:lambda:path/2015-03-31/functions/arn:aws:lambda:eu-central-1:196241772369:function:api_handler/invocations",
               "responses":{
                  "default":{
                     "statusCode":"200",
                     "responseParameters":{
                        "method.response.header.Access-Control-Allow-Origin":"'*'"
                     }
                  }
               },
               "passthroughBehavior":"when_no_match",
               "type":"aws_proxy"
            }
         },
         "post":{
            "summary":"Operations with tables",
            "description":"Operations with tables",
            "tags":[
               "Tables"
            ],
            "requestBody":{
               "description":"Create a table",
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/Table"
                     }
                  }
               }
            },
            "responses":{
               "200":{
                  "description":"200 response",
                  "headers":{
                     "Access-Control-Allow-Origin":{
                        "schema":{
                           "type":"string"
                        }
                     }
                  },
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/Empty"
                        }
                     }
                  }
               }
            },
            "security":[
               {
                  "authorizer":[

                  ]
               }
            ],
            "x-amazon-apigateway-integration":{
               "httpMethod":"POST",
               "uri":"arn:aws:apigateway:eu-central-1:lambda:path/2015-03-31/functions/arn:aws:lambda:eu-central-1:196241772369:function:api_handler/invocations",
               "responses":{
                  "default":{
                     "statusCode":"200",
                     "responseParameters":{
                        "method.response.header.Access-Control-Allow-Origin":"'*'"
                     }
                  }
               },
               "passthroughBehavior":"when_no_match",
               "type":"aws_proxy"
            }
         },
         "options":{
            "tags":[
               "Tables"
            ],
            "responses":{
               "200":{
                  "description":"200 response",
                  "headers":{
                     "Access-Control-Allow-Origin":{
                        "schema":{
                           "type":"string"
                        }
                     },
                     "Access-Control-Allow-Methods":{
                        "schema":{
                           "type":"string"
                        }
                     },
                     "Access-Control-Allow-Headers":{
                        "schema":{
                           "type":"string"
                        }
                     }
                  },
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/Empty"
                        }
                     }
                  }
               }
            },
            "x-amazon-apigateway-integration":{
               "responses":{
                  "default":{
                     "statusCode":"200",
                     "responseParameters":{
                        "method.response.header.Access-Control-Allow-Methods":"'*'",
                        "method.response.header.Access-Control-Allow-Headers":"'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token'",
                        "method.response.header.Access-Control-Allow-Origin":"'*'"
                     }
                  }
               },
               "requestTemplates":{
                  "application/json":"{\"statusCode\": 200}"
               },
               "passthroughBehavior":"when_no_match",
               "type":"mock"
            }
         }
      },
      "/signin":{
         "post":{
            "summary":"Operations with authentication",
            "description":"Operations with authentication",
            "tags":[
               "Authentication"
            ],
            "requestBody":{
               "description":"Create a table",
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/SignIn"
                     }
                  }
               }
            },
            "responses":{
               "200":{
                  "description":"200 response",
                  "headers":{
                     "Access-Control-Allow-Origin":{
                        "schema":{
                           "type":"string"
                        }
                     }
                  },
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/Empty"
                        }
                     }
                  }
               }
            },
            "x-amazon-apigateway-integration":{
               "httpMethod":"POST",
               "uri":"arn:aws:apigateway:eu-central-1:lambda:path/2015-03-31/functions/arn:aws:lambda:eu-central-1:196241772369:function:api_handler/invocations",
               "responses":{
                  "default":{
                     "statusCode":"200",
                     "responseParameters":{
                        "method.response.header.Access-Control-Allow-Origin":"'*'"
                     }
                  }
               },
               "passthroughBehavior":"when_no_match",
               "type":"aws_proxy"
            }
         },
         "options":{
            "tags":[
               "Authentication"
            ],
            "responses":{
               "200":{
                  "description":"200 response",
                  "headers":{
                     "Access-Control-Allow-Origin":{
                        "schema":{
                           "type":"string"
                        }
                     },
                     "Access-Control-Allow-Methods":{
                        "schema":{
                           "type":"string"
                        }
                     },
                     "Access-Control-Allow-Headers":{
                        "schema":{
                           "type":"string"
                        }
                     }
                  },
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/Empty"
                        }
                     }
                  }
               }
            },
            "x-amazon-apigateway-integration":{
               "responses":{
                  "default":{
                     "statusCode":"200",
                     "responseParameters":{
                        "method.response.header.Access-Control-Allow-Methods":"'*'",
                        "method.response.header.Access-Control-Allow-Headers":"'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token'",
                        "method.response.header.Access-Control-Allow-Origin":"'*'"
                     }
                  }
               },
               "requestTemplates":{
                  "application/json":"{\"statusCode\": 200}"
               },
               "passthroughBehavior":"when_no_match",
               "type":"mock"
            }
         }
      },
      "/signup":{
         "post":{
            "tags":[
               "Authentication"
            ],
            "summary":"Operations with authentication",
            "description":"Operations with authentication",
            "requestBody":{
               "description":"Create a new pet in the store",
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/UserCreate"
                     }
                  }
               },
               "required":true
            },
            "responses":{
               "200":{
                  "description":"200 response",
                  "headers":{
                     "Access-Control-Allow-Origin":{
                        "schema":{
                           "type":"string"
                        }
                     }
                  },
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/Empty"
                        }
                     }
                  }
               }
            },
            "x-amazon-apigateway-integration":{
               "httpMethod":"POST",
               "uri":"arn:aws:apigateway:eu-central-1:lambda:path/2015-03-31/functions/arn:aws:lambda:eu-central-1:196241772369:function:api_handler/invocations",
               "responses":{
                  "default":{
                     "statusCode":"200",
                     "responseParameters":{
                        "method.response.header.Access-Control-Allow-Origin":"'*'"
                     }
                  }
               },
               "passthroughBehavior":"when_no_match",
               "type":"aws_proxy"
            }
         },
         "options":{
            "tags":[
               "Authentication"
            ],
            "responses":{
               "200":{
                  "description":"200 response",
                  "headers":{
                     "Access-Control-Allow-Origin":{
                        "schema":{
                           "type":"string"
                        }
                     },
                     "Access-Control-Allow-Methods":{
                        "schema":{
                           "type":"string"
                        }
                     },
                     "Access-Control-Allow-Headers":{
                        "schema":{
                           "type":"string"
                        }
                     }
                  },
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/Empty"
                        }
                     }
                  }
               }
            },
            "x-amazon-apigateway-integration":{
               "responses":{
                  "default":{
                     "statusCode":"200",
                     "responseParameters":{
                        "method.response.header.Access-Control-Allow-Methods":"'*'",
                        "method.response.header.Access-Control-Allow-Headers":"'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token'",
                        "method.response.header.Access-Control-Allow-Origin":"'*'"
                     }
                  }
               },
               "requestTemplates":{
                  "application/json":"{\"statusCode\": 200}"
               },
               "passthroughBehavior":"when_no_match",
               "type":"mock"
            }
         }
      },
...
   },
   "components":{
      "schemas":{
         "SignIn":{
            "type":"object",
            "properties":{
               "email":{
                  "type":"string",
                  "example":"foobar@gmail.com"
               },
               "password":{
                  "type":"string",
                  "example":"VerySecret!pass$"
               }
            }
         },
         "Empty":{
            "title":"Empty Schema",
            "type":"object"
         },
         "Reservation":{
            "type":"object",
            "properties":{
               "tableNumber":{
                  "type":"integer",
                  "format":"int64",
                  "example":10
               },
               "clientName":{
                  "type":"string",
                  "example":"Test Clienr"
               },
               "phoneNumber":{
                  "type":"string",
                  "example":"2100900"
               },
               "date":{
                  "type":"string",
                  "example":"2024-06-30"
               },
               "slotTimeStart":{
                  "type":"string",
                  "example":"14:00"
               },
               "slotTimeEnd":{
                  "type":"string",
                  "example":"16:00"
               }
            }
         },
         "TableList":{
            "type":"object",
            "properties":{
               "tables":{
                  "type":"array",
                  "items":{
                     "$ref":"#/components/schemas/Table"
                  }
               }
            }
         },
         "Table":{
            "type":"object",
            "properties":{
               "places":{
                  "type":"integer",
                  "format":"int64",
                  "example":4
               },
               "minOrder":{
                  "type":"integer",
                  "format":"int64",
                  "example":10
               },
               "id":{
                  "type":"integer",
                  "format":"int64",
                  "example":10
               },
               "number":{
                  "type":"integer",
                  "format":"int64",
                  "example":101
               },
               "isVip":{
                  "type":"boolean",
                  "example":false
               }
            }
         },
         "ReservationsList":{
            "type":"object",
            "properties":{
               "reservations":{
                  "type":"array",
                  "items":{
                     "$ref":"#/components/schemas/Reservation"
                  }
               }
            }
         },
         "ReservationId":{
            "type":"object",
            "properties":{
               "reservationId":{
                  "type":"string",
                  "example":"37a694a6-803f-4999-b4ec-d18d56772f60"
               }
            }
         },
         "UserCreate":{
            "type":"object",
            "properties":{
               "firstName":{
                  "type":"string",
                  "example":"Foo"
               },
               "lastName":{
                  "type":"string",
                  "example":"Bar"
               },
               "email":{
                  "type":"string",
                  "example":"foobar@gmail.com"
               },
               "password":{
                  "type":"string",
                  "example":"VerySecret!pass$"
               }
            }
         }
      },
      "securitySchemes":{
         "authorizer":{
            "type":"apiKey",
            "name":"Authorization",
            "in":"header",
            "x-amazon-apigateway-authtype":"cognito_user_pools",
            "x-amazon-apigateway-authorizer":{
               "x-syndicate-cognito-userpool-names":[
                  "${booking_userpool}"
               ],
               "type":"cognito_user_pools"
            }
         }
      }
   }
}